add more to BatchExecShim
