+++ shims/common/src/main/scala/io/glutenproject/sql/shims/SparkShims.scala
@@ -38,20 +38,12 @@ case class SparkShimDescriptor(major: Int, minor: Int, patch: Int) extends ShimD
 trait SparkShims {
   def getShimDescriptor: ShimDescriptor
 
-  // https://issues.apache.org/jira/browse/SPARK-37377
-  def getKeyPartition(newPartitions: Seq[InputPartition], originalPartition: Partitioning)
-    : Seq[Seq[InputPartition]]
-
-  def newDatasourceRDD(sc: SparkContext, inputPartitions: Seq[Seq[InputPartition]],
-      partitionReaderFactory: PartitionReaderFactory,
-      columnarReads: Boolean,
-      customMetrics: Map[String, SQLMetric]): RDD[InternalRow]
-
   // for this purpose, change HashClusteredDistribution to ClusteredDistribution
   // https://github.com/apache/spark/pull/32875
   def getDistribution(leftKeys: Seq[Expression], rightKeys: Seq[Expression]): Seq[Distribution]
 
   // https://issues.apache.org/jira/browse/SPARK-36745
-  def applyPlan(plan: LogicalPlan, forceShuffledHashJoin: Boolean, backendLib: BackendLib): Seq[SparkPlan]
+  def applyPlan(plan: LogicalPlan, forceShuffledHashJoin: Boolean, backendLib: BackendLib)
+  : Seq[SparkPlan]
 
 }
