+++ shims/spark32/src/main/scala/io/glutenproject/sql/shims/spark32/Spark32Shims.scala
@@ -35,28 +35,6 @@ import org.apache.spark.sql.execution.SparkPlan
 class Spark32Shims extends SparkShims {
   override def getShimDescriptor: ShimDescriptor = SparkShimProvider.DESCRIPTOR
 
-  override def getKeyPartition(newPartitions: Seq[InputPartition],
-                      originalPartition: Partitioning): Seq[Seq[InputPartition]] = {
-    originalPartition match {
-      case p: DataSourcePartitioning if p.numPartitions != newPartitions.size =>
-        throw new SparkException(
-          "Data source must have preserved the original partitioning during runtime filtering; " +
-            s"reported num partitions: ${p.numPartitions}, " +
-            s"num partitions after runtime filtering: ${newPartitions.size}")
-      case _ =>
-      // no validation is needed as the data source did not report any specific partitioning
-    }
-    newPartitions.map(Seq(_))
-  }
-
-  override def newDatasourceRDD(sc: SparkContext, inputPartitions: Seq[Seq[InputPartition]],
-                                 partitionReaderFactory: PartitionReaderFactory,
-                                 columnarReads: Boolean,
-                                 customMetrics: Map[String, SQLMetric]): RDD[InternalRow] = {
-    new DataSourceRDD(sc, inputPartitions.flatten, partitionReaderFactory,
-      columnarReads, customMetrics)
-  }
-
   override def getDistribution(leftKeys: Seq[Expression], rightKeys: Seq[Expression])
     : Seq[Distribution] = {
     HashClusteredDistribution(leftKeys) :: HashClusteredDistribution(rightKeys) :: Nil
